Client
Les clients pourront envoyer des transactions via le server, pour garder une trace de l'envoie, une nouvelle transaction sera créée et stockée dans la liste des transactions du client à chaque envoie et reception.

Ce n'est pas encore fini, mais je commit quand même histroire que vous voyez un peu

+++ Wallet.java
@@ -42,7 +42,7 @@ public class Wallet {
 		return total;
 	}
 	
-	public Transaction send(String destinataire,float value ) {
+	public Transaction send(String destinataire,int value ) {
 		if(getSolde() < value) {
 			System.out.println("Fonds insuffisants. Transaction Refuse.");
 			return null;
